+++ tests/integrationtests/api/test_metadata_topics.py
@@ -1,7 +1,7 @@
 def test_metadata_topic(
     client,
     response_template,
-    metadata_topic_building_count,
+    metadata_topic_buildings,
 ):
     response = client.get("/metadata/topics")
     assert response.status_code == 200
@@ -9,21 +9,21 @@ def test_metadata_topic(
     content = response.json()
     result = content.pop("result")
     assert content == response_template
-    assert metadata_topic_building_count["building-count"] == result["building-count"]
+    assert metadata_topic_buildings["buildings"] == result["buildings"]
 
 
 def test_metadata_topic_by_key(
     client,
     response_template,
-    metadata_topic_building_count,
+    metadata_topic_buildings,
 ):
-    response = client.get("/metadata/topics/building-count")
+    response = client.get("/metadata/topics/buildings")
     assert response.status_code == 200
 
     content = response.json()
     result = content.pop("result")
     assert content == response_template
-    assert result == metadata_topic_building_count
+    assert result == metadata_topic_buildings
 
 
 def test_metadata_topic_by_key_not_found_error(client):