+++ tests/integrationtests/indicators/test_user_activity.py
@@ -11,10 +11,10 @@ pytestmark = pytest.mark.asyncio
 
 @oqapi_vcr.use_cassette
 async def test_preprocess(
-    topic_building_count,
+    topic_buildings,
     feature_germany_heidelberg,
 ):
-    indicator = UserActivity(topic_building_count, feature_germany_heidelberg)
+    indicator = UserActivity(topic_buildings, feature_germany_heidelberg)
     await indicator.preprocess()
     assert len(indicator.bin_total.users_abs) > 0
     assert isinstance(indicator.result.timestamp, datetime)
@@ -22,8 +22,8 @@ async def test_preprocess(
 
 
 @oqapi_vcr.use_cassette
-async def test_create_figure(topic_building_count, feature_germany_heidelberg):
-    indicator = UserActivity(topic_building_count, feature_germany_heidelberg)
+async def test_create_figure(topic_buildings, feature_germany_heidelberg):
+    indicator = UserActivity(topic_buildings, feature_germany_heidelberg)
     await indicator.preprocess()
     indicator.calculate()
     indicator.create_figure()